Rename windowing files to include windowing prefix
- Rename all windowing example files to include 'windowing' in their names:
  - r2c_c2r_legacy_callback_example.cu -> r2c_c2r_windowing_legacy_callback_example.cu
  - r2c_c2r_lto_callback_example.cpp -> r2c_c2r_windowing_lto_callback_example.cpp
  - r2c_c2r_lto_nvrtc_callback_example.cpp -> r2c_c2r_windowing_lto_nvrtc_callback_example.cpp
  - r2c_c2r_lto_callback_device.cu -> r2c_c2r_windowing_lto_callback_device.cu
  - r2c_c2r_reference.* -> r2c_c2r_windowing_reference.*
- Update CMakeLists.txt, README.md, and includes to use new names
- Rename executables to include windowing prefix
