+++ cuFFT/lto_callback_window_1d/src/r2c_c2r_windowing_lto_callback_device.cu
